--- app/filters/data.ts
@@ -1,5 +1,0 @@
-export const FILTER_DATA = {
-  genders: ['Все', 'Женский', 'Мужской'],
-  belts: ['10 гып', '9 гып', '8 гып', '7 гып', '6 гып', '5 гып', '4 гып', '3 гып', '2 гып', '1 гып', '1 дан', '2 дан', '3 дан', '4 дан', '5 дан', '6 дан', '7 дан', '8 дан', '9 дан'],
-  weights: ['до 30 кг', '30-40 кг', '40-50 кг', '50-60 кг', '60-70 кг', 'Более 70 кг']
-};
