+++ app/screens/SportsmanScreen.tsx
@@ -2,17 +2,16 @@ import { Ionicons } from '@expo/vector-icons';
 import { Stack, useLocalSearchParams, useRouter } from 'expo-router';
 import { useState } from 'react';
 import {
-    SafeAreaView,
-    ScrollView,
-    StatusBar,
-    Text,
-    TextInput,
-    TouchableOpacity,
-    View
+  SafeAreaView,
+  ScrollView,
+  StatusBar,
+  Text,
+  TextInput,
+  TouchableOpacity,
+  View
 } from 'react-native';
 
-import { styles } from './SportsmanScreen.styles';
-import { ATHLETES } from './data';
+import { StyleSheet } from 'react-native';
 
 export default function SportsmanScreen() {
   const router = useRouter();
@@ -64,7 +63,7 @@ const filteredAthletes = ATHLETES.filter(athlete => {
 
       <View style={styles.searchContainer}>
         {/* фильтр*/}
-        <TouchableOpacity onPress={() => router.push('/filters/FiltersScreen')}>
+        <TouchableOpacity onPress={() => router.push('/screens/FiltersScreen')}>
           <Ionicons name="funnel-outline" size={28} color="#fff" />
         </TouchableOpacity>
         
@@ -93,7 +92,7 @@ const filteredAthletes = ATHLETES.filter(athlete => {
               key={item.id} 
               style={styles.card}
               onPress={() => router.push({
-                pathname: '/sportsman-edit/SportsmanEditScreen',
+                pathname: '/screens/SportsmanEditScreen',
                 params: { id: item.id }
               })}
             >
@@ -114,10 +113,112 @@ const filteredAthletes = ATHLETES.filter(athlete => {
 
       <TouchableOpacity 
         style={styles.addButton}
-        onPress={() => router.push('/sportsman-edit/SportsmanEditScreen')}
+        onPress={() => router.push('/screens/SportsmanEditScreen')}
       >
         <Text style={styles.addButtonText}>Добавить спортсмена</Text>
       </TouchableOpacity>
     </SafeAreaView>
   );
 }
+
+
+
+export const styles = StyleSheet.create({
+  container: { 
+    flex: 1, 
+    backgroundColor: '#053552' 
+  },
+  header: { 
+    flexDirection: 'row', 
+    alignItems: 'center', 
+    paddingHorizontal: 20, 
+    paddingVertical: 15 
+  },
+  headerTitleRow: { 
+    flexDirection: 'row', 
+    alignItems: 'center', 
+    marginLeft: 15 
+  },
+  headerTitle: { 
+    color: '#fff', 
+    fontSize: 24, 
+    fontWeight: 'bold', 
+    marginLeft: 8 
+  },
+  searchContainer: { 
+    flexDirection: 'row', 
+    alignItems: 'center', 
+    paddingHorizontal: 20, 
+    marginBottom: 20,
+    gap: 15
+  },
+  searchInputWrapper: { 
+    flex: 1, 
+    backgroundColor: '#fff', 
+    borderRadius: 4, 
+    height: 40, 
+    justifyContent: 'center',
+    paddingHorizontal: 10
+  },
+  searchInput: { 
+    color: '#000', 
+    fontSize: 16 
+  },
+  listContent: { 
+    paddingHorizontal: 20, 
+    paddingBottom: 100 
+  },
+  card: { 
+    backgroundColor: '#0E4E75', 
+    borderRadius: 8, 
+    padding: 15, 
+    marginBottom: 15,
+    // Используем современный способ для теней
+    boxShadow: '0px 4px 4px rgba(0, 0, 0, 0.25)', 
+  },
+  cardName: { 
+    color: '#fff', 
+    fontSize: 18, 
+    fontWeight: 'bold', 
+    marginBottom: 5 
+  },
+  cardInfoRow: { 
+    flexDirection: 'row', 
+    alignItems: 'center', 
+    marginBottom: 3 
+  },
+  statusDot: { 
+    width: 14, 
+    height: 14, 
+    borderRadius: 7, 
+    marginRight: 8 
+  },
+  cardInfoText: { 
+    color: '#fff', 
+    fontSize: 16 
+  },
+  addButton: {
+    position: 'absolute',
+    bottom: 30,
+    left: 60,
+    right: 60,
+    backgroundColor: '#084366',
+    paddingVertical: 15,
+    borderRadius: 30,
+    alignItems: 'center',
+    borderWidth: 1,
+    borderColor: '#1A6B9B',
+    boxShadow: '0px 4px 10px rgba(0, 0, 0, 0.3)',
+  },
+  addButtonText: { 
+    color: '#fff', 
+    fontSize: 16, 
+    fontWeight: '600' 
+  }
+});
+
+export const ATHLETES = [
+  { id: '1', name: 'Иван Петров', age: 15, belt: '1 дан', gender: 'Мужской', weight: 45, statusColor: '#FF4141' },
+  { id: '2', name: 'Алексей Сидоров', age: 25, belt: '10 гып', gender: 'Мужской', weight: 60, statusColor: '#4CAF50' },
+  { id: '3', name: 'Мария Иванова', age: 14, belt: '5 гып', gender: 'Женский', weight: 30, statusColor: '#FF9800' },
+];
