--- app/sportsman/data.ts
@@ -1,5 +1,0 @@
-export const ATHLETES = [
-  { id: '1', name: 'Иван Петров', age: 15, belt: '1 дан', gender: 'Мужской', weight: 45, statusColor: '#FF4141' },
-  { id: '2', name: 'Алексей Сидоров', age: 25, belt: '10 гып', gender: 'Мужской', weight: 60, statusColor: '#4CAF50' },
-  { id: '3', name: 'Мария Иванова', age: 14, belt: '5 гып', gender: 'Женский', weight: 30, statusColor: '#FF9800' },
-];
